// src/app/modules/setup/models/crud/session.ts
export class Session {
    SESSION_ID: string;
    ORG_CODE: string;
    SESSION_CODE: string;
    SESSION_NAME: string;
    SESSION_NAME_BANGLA: string;
    START_DATE: string;
    END_DATE: string;
    SESSION_DTL_ID: number[];
    BILL_MONTH: any[];
    DTL_START_DATE: string[];
    DTL_END_DATE: string[];
    BILL_DATE: string[];
    DUE_DATE: string[];
    Row_Status: number[];
    User_Code: string;
}

// BILL_DATE: (12) ['2023-07-01T00:00:00', '2023-08-01T00:00:00', '2023-09-01T00:00:00', '2023-10-01T00:00:00', '2023-11-01T00:00:00', '2023-12-01T00:00:00', '2024-01-01T00:00:00', '2024-02-01T00:00:00', '2024-03-01T00:00:00', '2024-04-01T00:00:00', '2024-05-01T00:00:00', '2024-06-01T00:00:00']
// BILL_MONTH: (12) ['202307', '202308', '202309', '202310', '202311', '202312', '202401', '202402', '202403', '202404', '202405', '202406']
// DTL_END_DATE: (12) ['2023-07-01T00:00:00', '2023-08-01T00:00:00', '2023-09-01T00:00:00', '2023-10-01T00:00:00', '2023-11-01T00:00:00', '2023-12-01T00:00:00', '2024-01-01T00:00:00', '2024-02-01T00:00:00', '2024-03-01T00:00:00', '2024-04-01T00:00:00', '2024-05-01T00:00:00', '2024-06-01T00:00:00']
// DTL_START_DATE: (12) ['2023-07-01T00:00:00', '2023-08-01T00:00:00', '2023-09-01T00:00:00', '2023-10-01T00:00:00', '2023-11-01T00:00:00', '2023-12-01T00:00:00', '2024-01-01T00:00:00', '2024-02-01T00:00:00', '2024-03-01T00:00:00', '2024-04-01T00:00:00', '2024-05-01T00:00:00', '2024-06-01T00:00:00']
// DUE_DATE: (12) ['2023-07-15T00:00:00', '2023-08-15T00:00:00', '2023-09-15T00:00:00', '2023-10-15T00:00:00', '2023-11-15T00:00:00', '2023-12-15T00:00:00', '2024-01-15T00:00:00', '2024-02-15T00:00:00', '2024-03-15T00:00:00', '2024-04-15T00:00:00', '2024-05-15T00:00:00', '2024-06-15T00:00:00']
// END_DATE: "2024-06-30T00:00:00"
// ORG_CODE: "00000146"
// SESSION_CODE: "00000112"
// SESSION_DTL_ID: (12) [620, 621, 622, 623, 624, 625, 626, 627, 628, 629, 630, 631]
// SESSION_ID: 112
// SESSION_NAME: "2023-24"
// START_DATE: "2023-07-01T00:00:00"